moved checkout outside of user subroute
