+++ src/screens/App/screens/Checkout/components/PaymentDetails.jsx
@@ -1,8 +1,8 @@
 import React from "react";
-import useCheckout from "../../../../../../../hooks/useCheckout";
-import visaIcon from "../../../../../../../shared/images/visa.png";
-import mastercardIcon from "../../../../../../../shared/images/mastercard.png";
-import paypalIcon from "../../../../../../../shared/images/paypal.png";
+import useCheckout from "../../../../../hooks/useCheckout";
+import visaIcon from "../../../../../shared/images/visa.png";
+import mastercardIcon from "../../../../../shared/images/mastercard.png";
+import paypalIcon from "../../../../../shared/images/paypal.png";
 import PMBody from "./paymentDetails/MethodBody";
 import CashOnDelivery from "./paymentDetails/CashOnDelivery";
 import CreditCard from "./paymentDetails/CreditCard";
