+++ src/screens/App/screens/Checkout/components/paymentDetails/CashOnDelivery.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import useCheckout from "../../../../../../../../hooks/useCheckout";
+import useCheckout from "../../../../../../hooks/useCheckout";
 import { PaymentCTA } from "../utils/CallToAction";
 
 function CashOnDelivery() {
