+++ src/screens/App/screens/Checkout/components/paymentDetails/MethodBody.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import useCheckout from "../../../../../../../../hooks/useCheckout";
+import useCheckout from "../../../../../../hooks/useCheckout";
 
 function PaymentMethodBody({ id, children, active }) {
   const { togglePaymentOption } = useCheckout();
