+++ src/screens/App/screens/Checkout/index.js
