+++ src/screens/App/screens/Checkout/route.js
@@ -1,0 +1,10 @@
+import routes from "../../../../shared/Route/routes";
+
+import Checkout from ".";
+
+const route = {
+  component: Checkout,
+  path: routes.CHECKOUT.path,
+  exact: true,
+};
+export default route;
