--- src/screens/App/screens/User/screens/Checkout/route.js
@@ -1,8 +1,0 @@
-import routes from "../../../../../../shared/Route/routes";
-import Checkout from ".";
-
-const route = {
-  path: routes.USER.subroutes.CHECKOUT.path,
-  component: Checkout,
-};
-export default route;
